+++ app/config/routing.yml
@@ -47,9 +47,6 @@ oro_default:
         template:    PimEnrichBundle::index.html.twig
         _controller: FrameworkBundle:Template:template
 
-oro_user:
-    resource: "@OroUserBundle/Resources/config/oro/routing.yml"
-
 oro_config:
     resource: "@OroConfigBundle/Resources/config/oro/routing.yml"
