+++ src/Pim/Bundle/UserBundle/Controller/Rest/SecurityController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\UserBundle\Controller;
+namespace Pim\Bundle\UserBundle\Controller\Rest;
 
 use Oro\Bundle\SecurityBundle\Metadata\AclAnnotationProvider;
 use Oro\Bundle\SecurityBundle\SecurityFacade;
@@ -13,7 +13,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class SecurityRestController
+class SecurityController
 {
     /** @var SecurityFacade */
     protected $securityFacade;
