+++ src/Pim/Bundle/UserBundle/Controller/Rest/UserController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\UserBundle\Controller;
+namespace Pim\Bundle\UserBundle\Controller\Rest;
 
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -15,7 +15,7 @@ use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class UserRestController
+class UserController
 {
     /** @var TokenStorageInterface */
     protected $tokenStorage;
