+++ src/Pim/Bundle/UserBundle/Controller/Rest/UserGroupController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\UserBundle\Controller;
+namespace Pim\Bundle\UserBundle\Controller\Rest;
 
 use Pim\Bundle\UserBundle\Doctrine\ORM\Repository\GroupRepository;
 use Pim\Component\User\Model\GroupInterface;
@@ -13,7 +13,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class UserGroupRestController
+class UserGroupController
 {
     /** @var GroupRepository */
     protected $groupRepository;
