+++ src/Pim/Bundle/UserBundle/Resources/config/controllers.yml
@@ -1,7 +1,7 @@
 parameters:
-    pim_user.controller.user_rest.class: Pim\Bundle\UserBundle\Controller\UserRestController
-    pim_user.controller.security_rest.class: Pim\Bundle\UserBundle\Controller\SecurityRestController
-    pim_user.controller.user_group_rest.class: Pim\Bundle\UserBundle\Controller\UserGroupRestController
+    pim_user.controller.user_rest.class: Pim\Bundle\UserBundle\Controller\Rest\UserController
+    pim_user.controller.security_rest.class: Pim\Bundle\UserBundle\Controller\Rest\SecurityController
+    pim_user.controller.user_group_rest.class: Pim\Bundle\UserBundle\Controller\Rest\UserGroupController
 
 services:
     pim_user.controller.user_rest:
